wolf_sheep: Merge benchmark model into example model
Key changes and improvements in this merged version:

1. Used the experimental cell space features:
   - Switched to the more efficient `OrthogonalVonNeumannGrid`
   - Leveraged `CellAgent` and `FixedAgent` base classes
   - Used the cell's neighborhood property for movement

2. Implemented discrete event scheduling for grass regrowth:
   - Used the experimental `ABMSimulator` for event scheduling
   - Replaced the countdown-based grass regrowth with scheduled events

3. Modern Mesa practices:
   - Used `agents_by_type` for agent management
   - Employed `shuffle_do()` for efficient random activation
   - Proper initialization of the Model class with `super().__init__()`
   - Clear data collection setup with dedicated model reporters

4. Code organization:
   - Maintained separation between agents and model
   - Added detailed docstrings
   - Used property decorators for grass state management
   - Consistent style and naming conventions

5. Additional improvements:
   - Made grass optional while keeping full functionality
   - Improved type hints and property usage
   - More efficient agent selection and movement
   - Better encapsulation of agent behaviors
