--- benchmarks/WolfSheep/__init__.py
@@ -1,1 +1,0 @@
-"""Wolf-Sheep Predation Model for performance benchmarking."""
